refactor: 내가 남긴 리뷰 페이지 section 영역 스타일 추가
